Move entries from Gruntfile to webpackEntries.js
